test reward distribution in withdraw_inactive_stake
