+++ program/tests/tests/mod.rs
@@ -5,7 +5,6 @@
 
 pub mod add_remove_validator;
 pub mod change_reward_distribution;
-// pub mod collect_validator_fee;
 pub mod deposit;
 pub mod limits;
 pub mod maintainers;
